Host page and reformating
Created the host pages for the VanLife for those required pages. Added the imports and initial layout to connect everything together.

Reshuffled files to be in the correct places i.e., `Vans.jsx` and `VanDetails.jsx` are to be in a folder of their own. Also, removed unused variable for the time being.